refactor: rename files ibm -> index, split utils
