+++ libs/providers/langchain-ibm/src/chat_models/tests/index.test.ts
@@ -16,7 +16,7 @@ import {
   ChatWatsonxGatewayInput,
   ChatWatsonxInput,
   WatsonxCallParams,
-} from "../ibm.js";
+} from "../index.js";
 import { WatsonxAuth } from "../../types.js";
 
 const fakeAuthProp = {
