+++ libs/providers/langchain-ibm/src/embeddings/index.ts
@@ -10,11 +10,8 @@ import {
   WatsonxConfigurationError,
   WatsonxUnsupportedOperationError,
 } from "../types.js";
-import {
-  checkValidProps,
-  expectOneOf,
-  initWatsonxOrGatewayInstance,
-} from "../utils/ibm.js";
+import { checkValidProps, expectOneOf } from "../utils/validation.js";
+import { initWatsonxOrGatewayInstance } from "../utils/instance.js";
 
 export interface WatsonxEmbeddingsParams
   extends
