+++ libs/providers/langchain-ibm/src/embeddings/tests/index.test.ts
@@ -1,8 +1,8 @@
 /* oxlint-disable dot-notation */
 import { Gateway } from "@ibm-cloud/watsonx-ai/gateway";
 import { vi, expect, describe, test } from "vitest";
-import { WatsonxEmbeddings, WatsonxInputEmbeddings } from "../ibm.js";
-import { testProperties } from "../../llms/tests/ibm.test.js";
+import { WatsonxEmbeddings, WatsonxInputEmbeddings } from "../index.js";
+import { testProperties } from "../../llms/tests/index.test.js";
 
 const fakeAuthProp = {
   watsonxAIAuthType: "iam",
