--- libs/providers/langchain-ibm/src/utils/ibm.ts
@@ -1,286 +1,0 @@
-/* oxlint-disable @typescript-eslint/no-explicit-any */
-import { WatsonXAI } from "@ibm-cloud/watsonx-ai";
-import {
-  JsonOutputKeyToolsParserParamsInterop,
-  JsonOutputToolsParser,
-} from "@langchain/core/output_parsers/openai_tools";
-import { OutputParserException } from "@langchain/core/output_parsers";
-import { z } from "zod/v3";
-import { ChatGeneration } from "@langchain/core/outputs";
-import { AIMessageChunk } from "@langchain/core/messages";
-import { ToolCall } from "@langchain/core/messages/tool";
-import {
-  InteropZodType,
-  interopSafeParseAsync,
-} from "@langchain/core/utils/types";
-import { Gateway } from "@ibm-cloud/watsonx-ai/gateway";
-import {
-  WatsonxAuth,
-  WatsonxInit,
-  WatsonxAuthenticationError,
-  WatsonxValidationError,
-  WatsonxUnsupportedOperationError,
-} from "../types.js";
-import { prepareInstanceConfig } from "../auth/index.js";
-
-/**
- * Initializes and returns a WatsonX AI or Gateway instance with authentication.
- * 
- * @param params - Initialization and authentication parameters
- * @param useGateway - If true, returns Gateway instance; otherwise returns WatsonXAI instance
- * @returns Configured WatsonXAI or Gateway instance
- */
-export function initWatsonxOrGatewayInstance(
-  params: WatsonxAuth & Omit<WatsonxInit, "authenticator">,
-  useGateway: true,
-): Gateway;
-export function initWatsonxOrGatewayInstance(
-  params: WatsonxAuth & Omit<WatsonxInit, "authenticator">,
-  useGateway?: false,
-): WatsonXAI;
-export function initWatsonxOrGatewayInstance(
-  params: WatsonxAuth & Omit<WatsonxInit, "authenticator">,
-  useGateway = false,
-): WatsonXAI | Gateway {
-  const config = prepareInstanceConfig(params);
-  try {
-    return useGateway ? new Gateway(config) : new WatsonXAI(config);
-  } catch (e) {
-    throw new WatsonxAuthenticationError(
-      "You have not provided any type of authentication",
-    );
-  }
-}
-
-const TOOL_CALL_ID_PATTERN = /^[a-zA-Z0-9]{9}$/;
-
-export function _isValidMistralToolCallId(toolCallId: string): boolean {
-  return TOOL_CALL_ID_PATTERN.test(toolCallId);
-}
-
-function _base62Encode(num: number): string {
-  let numCopy = num;
-  const base62 =
-    "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
-  if (numCopy === 0) return base62[0];
-  const arr: string[] = [];
-  const base = base62.length;
-  while (numCopy) {
-    arr.push(base62[numCopy % base]);
-    numCopy = Math.floor(numCopy / base);
-  }
-  return arr.reverse().join("");
-}
-
-function _simpleHash(str: string): number {
-  let hash = 0;
-  for (let i = 0; i < str.length; i += 1) {
-    const char = str.charCodeAt(i);
-    hash = (hash << 5) - hash + char;
-    hash &= hash;
-  }
-  return Math.abs(hash);
-}
-
-export function _convertToolCallIdToMistralCompatible(
-  toolCallId: string
-): string {
-  if (_isValidMistralToolCallId(toolCallId)) {
-    return toolCallId;
-  } else {
-    const hash = _simpleHash(toolCallId);
-    const base62Str = _base62Encode(hash);
-    if (base62Str.length >= 9) {
-      return base62Str.slice(0, 9);
-    } else {
-      return base62Str.padStart(9, "0");
-    }
-  }
-}
-
-interface WatsonxToolsOutputParserParams<
-  T extends Record<string, any>,
-> extends JsonOutputKeyToolsParserParamsInterop<T> {}
-
-export class WatsonxToolsOutputParser<
-  T extends Record<string, any> = Record<string, any>,
-> extends JsonOutputToolsParser<T> {
-  static lc_name() {
-    return "WatsonxToolsOutputParser";
-  }
-
-  lc_namespace = ["langchain", "watsonx", "output_parsers"];
-
-  returnId = false;
-
-  keyName: string;
-
-  returnSingle = false;
-
-  zodSchema?: InteropZodType<T>;
-
-  latestCorrect?: ToolCall;
-
-  constructor(params: WatsonxToolsOutputParserParams<T>) {
-    super(params);
-    this.keyName = params.keyName;
-    this.returnSingle = params.returnSingle ?? this.returnSingle;
-    this.zodSchema = params.zodSchema;
-  }
-
-  protected async _validateResult(result: unknown): Promise<T> {
-    let parsedResult = result;
-    if (typeof result === "string") {
-      try {
-        parsedResult = JSON.parse(result);
-      } catch (e: any) {
-        throw new OutputParserException(
-          `Failed to parse. Text: "${JSON.stringify(
-            result,
-            null,
-            2
-          )}". Error: ${JSON.stringify(e.message)}`,
-          result
-        );
-      }
-    } else {
-      parsedResult = result;
-    }
-    if (this.zodSchema === undefined) {
-      return parsedResult as T;
-    }
-    const zodParsedResult = await interopSafeParseAsync(
-      this.zodSchema,
-      parsedResult
-    );
-    if (zodParsedResult.success) {
-      return zodParsedResult.data;
-    } else {
-      throw new OutputParserException(
-        `Failed to parse. Text: "${JSON.stringify(
-          result,
-          null,
-          2
-        )}". Error: ${JSON.stringify(zodParsedResult.error.issues)}`,
-        JSON.stringify(result, null, 2)
-      );
-    }
-  }
-
-  async parsePartialResult(generations: ChatGeneration[]): Promise<T> {
-    const tools = generations.flatMap((generation) => {
-      const message = generation.message as AIMessageChunk;
-      if (!Array.isArray(message.tool_calls)) {
-        return [];
-      }
-      const tool = message.tool_calls;
-      return tool;
-    });
-
-    if (tools[0] === undefined) {
-      if (this.latestCorrect) {
-        tools.push(this.latestCorrect);
-      } else {
-        const toolCall: ToolCall = { name: "", args: {} };
-        tools.push(toolCall);
-      }
-    }
-
-    const [tool] = tools;
-    tool.name = "";
-    this.latestCorrect = tool;
-    return tool.args as T;
-  }
-}
-
-export function jsonSchemaToZod(obj: WatsonXAI.JsonObject | undefined) {
-  if (obj?.properties && obj.type === "object") {
-    const shape: Record<string, any> = {};
-
-    Object.keys(obj.properties).forEach((key) => {
-      if (obj.properties) {
-        const prop = obj.properties[key];
-
-        let zodType;
-        if (prop.type === "string") {
-          zodType = z.string();
-          if (prop?.pattern) {
-            zodType = zodType.regex(prop.pattern, "Invalid pattern");
-          }
-        } else if (
-          prop.type === "number" ||
-          prop.type === "integer" ||
-          prop.type === "float"
-        ) {
-          zodType = z.number();
-          if (typeof prop?.minimum === "number") {
-            zodType = zodType.min(prop.minimum, {
-              message: `${key} must be at least ${prop.minimum}`,
-            });
-          }
-          if (prop?.maximum)
-            zodType = zodType.lte(prop.maximum, {
-              message: `${key} must be maximum of ${prop.maximum}`,
-            });
-        } else if (prop.type === "boolean") zodType = z.boolean();
-        else if (prop.type === "array")
-          zodType = z.array(
-            prop.items ? jsonSchemaToZod(prop.items) : z.string(),
-          );
-        else if (prop.type === "object") {
-          zodType = jsonSchemaToZod(prop);
-        } else
-          throw new WatsonxUnsupportedOperationError(
-            `Unsupported type: ${prop.type}`,
-          );
-
-        if (prop.description) {
-          zodType = zodType.describe(prop.description);
-        }
-
-        if (!obj.required?.includes(key)) {
-          zodType = zodType.optional();
-        }
-
-        shape[key] = zodType;
-      }
-    });
-    return z.object(shape);
-  }
-  throw new WatsonxUnsupportedOperationError("Unsupported root schema type");
-}
-
-export const expectOneOf = (
-  params: Record<string, any>,
-  keys: string[],
-  exactlyOneOf = false
-) => {
-  const provided = keys.filter(
-    (key) => key in params && params[key] !== undefined
-  );
-  if (exactlyOneOf && provided.length !== 1) {
-    throw new WatsonxValidationError(
-      `Expected exactly one of: ${keys.join(", ")}. Got: ${provided.join(", ")}`,
-    );
-  } else if (!exactlyOneOf && provided.length > 1) {
-    throw new WatsonxValidationError(
-      `Expected one of: ${keys.join(", ")} or none. Got: ${provided.join(", ")}`,
-    );
-  }
-};
-
-export const checkValidProps = (
-  params: Record<string, any>,
-  allowedKeys: string[]
-) => {
-  const unexpected = Object.keys(params).filter(
-    (key) => !allowedKeys.includes(key)
-  );
-  if (unexpected.length > 0) {
-    throw new WatsonxValidationError(
-      `Unexpected properties: ${unexpected.join(
-        ", ",
-      )}. Expected only: ${allowedKeys.join(", ")}.`,
-    );
-  }
-};
